Add files via upload
During the second year Second semester of my course program I developed this simple side for my end semester project. This is a simple web application for a computer and IT related service provider. This will have the basic information of the service provides, their services, their shops and the details of their business. It 4 different pages for Home, About us , services and Register. Using the register form users can register for services which they wanted to get from the service provide. However due some of the errors in the PHP and XAMPP the database in not connecting properly. Other than the database the page works properly.